+++ keras_nlp/src/models/pali_gemma/pali_gemma_vit_test.py
@@ -14,14 +14,16 @@
 import numpy as np
 import pytest
 
-from keras_nlp.src.models.paligemma.vision_embeddings import VisionEmbeddings
-from keras_nlp.src.models.paligemma.vit import PaliGemmaViT
-from keras_nlp.src.models.paligemma.vit import VitEncoder
+from keras_nlp.src.models.pali_gemma.pali_gemma_vit import PaliGemmaVit
+from keras_nlp.src.models.pali_gemma.pali_gemma_vit import (
+    PaliGemmaVitEmbeddings,
+)
+from keras_nlp.src.models.pali_gemma.pali_gemma_vit import PaliGemmaVitEncoder
 from keras_nlp.src.tests.test_case import TestCase
 
 
 @pytest.mark.keras_3_only
-class VITTest(TestCase):
+class PaliGemmaVitTest(TestCase):
     def test_vit_encoder(self):
         #  encoder calls Attention and CLIPLayer, both of which gets
         # verified with this test
@@ -31,7 +33,7 @@ class VITTest(TestCase):
         num_layers = 4
         heads = 8
         intermediate_dim = 256
-        vit_encoder = VitEncoder(
+        vit_encoder = PaliGemmaVitEncoder(
             hidden_dim,
             num_layers,
             heads,
@@ -46,13 +48,13 @@ class VITTest(TestCase):
         )
 
     def test_vision_embeddings(self):
-        embeddings_layer = VisionEmbeddings(3)
+        embeddings_layer = PaliGemmaVitEmbeddings(3)
         dummy_input = np.ones([1, 224, 224, 3])
         vision_embeddings = embeddings_layer(dummy_input)
         self.assertEqual(vision_embeddings.shape, (1, 256, 3))
 
     def test_vit_output_shape(self):
-        embeddings_layer = PaliGemmaViT(image_size=224)
+        embeddings_layer = PaliGemmaVit(image_size=224)
         dummy_input = np.ones([1, 224, 224, 3])
         image_embeddings = embeddings_layer(dummy_input)
         self.assertEqual(image_embeddings.shape, (1, 256, 2048))
